Implemented TracePath()
Not perfect yet. The render takes longer than it should, likely due to my heavy use of shared pointers, so I need to fix that, possibly refactor to switch them to unique pointers. Also anywhere that is supposed to be white light is actually cyan, idk why so that needs to be fixed.