Optimised RayHit and BRDF shared pointers
Instead of creating an empty material with a diffuse BRDF for missed rays, they are now stored as null, since they are never used. It's important that TracePath() returns before doing any calculations if the ray misses. This was causing performance issues because I was creating and destroying multiple shared pointers per ray.