Actually fixed the BVH
For non leaf boxes, I was returning left.Intersect() || right.Intersect(). If the left intersection returned true, the OR operation would ignore the right and just return the left value, since the right wouldn't have changed it. Fixed it by applying the left and right intersection to two bool variables and preforming the OR on them to ensure both boxes have their intersect functions called.